Add files via upload
These files were originally written to predict volume of property sales in the UK (using only open source, macro-level financial data). The uploaded versions are generic versions which can form the basis of using LGBM or XGBoost on other time series data. 

The files are WIP in that the ensemble has not been optimised.